[refactor] separate run-test logic from cli
